Cleanup for Recommended: IrqlSetTooHigh
