+++ src/drivers/general/queries/IrqlSetTooHigh/IrqlSetTooHigh.qhelp
@@ -1,0 +1,38 @@
+<!DOCTYPE qhelp PUBLIC "-//Semmle//qhelp//EN" "qhelp.dtd">
+<qhelp>
+	<overview>
+		<p>
+			The function has raised the IRQL to a level above what is allowed.
+		</p>
+	</overview>
+	<recommendation>
+		<p>
+			A function has been annotated as having a max IRQL, but the execution of that function raises the IRQL above that maximum.  If you have applied custom IRQL annotations to your own functions, confirm that they are accurate.
+		</p>
+	</recommendation>
+	<example>
+		<p>
+			In this example, the driver tries to raise the IRQL to HIGH_LEVEL while in a dispatch routine.  This should be avoided.
+		</p>
+		<sample language="c"><![CDATA[
+			// Within a dispatch routine
+			KeRaiseIrql(HIGH_LEVEL, &oldIRQL);
+			]]>
+		</sample>
+	</example>
+	<references>
+		<li>
+			<a href="https://learn.microsoft.com/en-us/windows-hardware/drivers/devtest/28150-function-causes-irq-level-to-be-set-above-max">
+				C28150
+			</a>
+			<a href="https://learn.microsoft.com/en-us/windows-hardware/drivers/devtest/irql-annotations-for-drivers">
+				IRQL annotations for drivers
+			</a>
+		</li>
+	</references>
+	<semmleNotes>
+		<p>This query uses interprocedural data-flow analysis and can take a large amount of CPU time and memory to run.</p>
+		<p>This query may provide false positives in cases where functions are not annotated with their expected IRQL ranges or behaviors.</p>
+		<p>For information on how to annotate your functions with information about how they adjust the IRQL, see "IRQL annotations for drivers" in the references section.</p>
+	</semmleNotes>
+</qhelp>
